+++ src/converter.ts
@@ -77,6 +77,7 @@ export interface ConverterOption {
   options: MarpitOptions
   output?: string | false
   pages?: boolean | number[]
+  parallel?: number
   pdfNotes?: boolean
   pdfOutlines?:
     | false
@@ -225,6 +226,8 @@ export class Converter {
     }
 
     if (!opts.onlyScanning) {
+      debug('Converting %s ...', file.relativePath())
+
       const files: File[] = []
       switch (this.options.type) {
         case ConvertType.pdf:
@@ -283,7 +286,24 @@ export class Converter {
     if (!inputDir && output && output !== '-' && files.length > 1)
       error('Output path cannot specify with processing multiple files.')
 
-    for (const file of files) await this.convertFile(file, opts)
+    const parallel = Math.max(1, this.options.parallel ?? 1)
+    const queue = [...files]
+
+    const workers = Array.from({ length: parallel }, async (_, i) => {
+      debug(`[Worker ${i + 1}] Start processing ...`)
+
+      let file: File | undefined
+
+      while ((file = queue.shift())) {
+        debug(`[Worker ${i + 1}] Processing ${file.absolutePath} ...`)
+        await this.convertFile(file, opts)
+      }
+
+      debug(`[Worker ${i + 1}] Finish processing.`)
+    })
+
+    await Promise.all(workers)
+    debug(`Batch processing has been completed.`)
   }
 
   private convertFileToHTML(tpl: TemplateResult, file: File): File {